fix(react-doctor): small state-pattern and circular-dependency fixes
- circular-dependency: extract IconWrapper + MessageActionItem into
  message-action-primitives.tsx; audio-player and massage-action both
  import from it, breaking the cycle. massage-action re-exports both
  for backward compat.
- no-mirror-prop-effect / no-reset-all-state-on-prop-change: add
  key={employee.id} and key={team.id} at the EmployeeEditor /
  TeamEditor call sites in workforce-page; remove the now-redundant
  useEffect(()=>setDraft(prop),[prop]) from each editor (React
  remounts on key change, so useState initializer always gets fresh value).
